CP-005-Login-new-logic: Fetched data stored to VueX (User Claims and User Loc)
